% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqgwas.R
\name{qqgwas}
\alias{qqgwas}
\title{QQ plot for > 1M points}
\usage{
qqgwas(pvalues, n = length(pvalues), pointsize = 4, color = "dodgerblue", ...)
}
\arguments{
\item{pvalues}{array of p-values}

\item{n}{number of tests performed. defaults to \code{length(pvalues)}}

\item{pointsize}{size of points}

\item{color}{color of points}

\item{...}{other arguments passed to \code{geom_scattermore()}}
}
\description{
Saving a QQ plot with >1M p-values can be slow and create a huge PDF since each point is saved when using \code{geom_point()}. Instead, use \code{geom_scattermore()} to rasterize points. This saves an image of the points, so overlapping points are just plotted once. This dramatically reduces the size of editable PDFs.
}
\details{
QQ plot for > 1M points using scattermore to create small editable images
}
\examples{
qqgwas(runif(1000000))

}
